pyproject.toml
[project]
name = "uv-fastapi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.112.1",
    "strawberry-graphql[fastapi]>=0.282.0",
]
